requirements.txt
flask~=3.1.0
pydantic~=2.11.3
flask-cors~=5.0.1
bcrypt~=4.3.0
mysql-connector-python~=9.3.0
PyJWT~=2.10.1
ultralytics~=8.3.124
pillow~=11.2.1